requirements.txt
click
pyserial
flask
gunicorn